+++ app/scripts/modules/kubernetes/src/pipelines/stages/deployManifest/manifestStatus/DeployStatus.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { get, upperFirst } from 'lodash';
+import { get } from 'lodash';
 import {
   IExecutionDetailsSectionProps,
   ExecutionDetailsSection,
@@ -10,26 +10,6 @@ import {
 import { KubernetesManifestService, IStageManifest } from '../../../../manifest/manifest.service';
 
 import { ManifestStatus } from './ManifestStatus';
-
-// from https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.12/
-const BUILT_IN_GROUPS = [
-  '',
-  'core',
-  'batch',
-  'apps',
-  'extensions',
-  'storage.k8s.io',
-  'apiextensions.k8s.io',
-  'apiregistration.k8s.io',
-  'policy',
-  'scheduling.k8s.io',
-  'settings.k8s.io',
-  'authorization.k8s.io',
-  'authentication.k8s.io',
-  'rbac.authorization.k8s.io',
-  'certifcates.k8s.io',
-  'networking.k8s.io',
-];
 
 export interface IManifestSubscription {
   id: string;
@@ -76,11 +56,7 @@ export class DeployStatus extends React.Component<IExecutionDetailsSectionProps,
   }
 
   private subscribeToManifestUpdates(id: string, manifest: IStageManifest): () => void {
-    const params = {
-      account: this.props.stage.context.account,
-      name: this.scopedKind(manifest) + ' ' + manifest.metadata.name,
-      location: manifest.metadata.namespace == null ? '_' : manifest.metadata.namespace,
-    };
+    const params = KubernetesManifestService.stageManifestToManifestParams(manifest, this.props.stage.context.account);
     return KubernetesManifestService.subscribe(this.props.application, params, (updated: IManifest) => {
       const idx = this.state.subscriptions.findIndex(sub => sub.id === id);
       if (idx !== -1) {
@@ -94,26 +70,6 @@ export class DeployStatus extends React.Component<IExecutionDetailsSectionProps,
 
   private unsubscribeAll() {
     this.state.subscriptions.forEach(({ unsubscribe }) => unsubscribe());
-  }
-
-  private apiGroup(manifest: IStageManifest): string {
-    const parts = (manifest.apiVersion || '_').split('/');
-    if (parts.length < 2) {
-      return '';
-    }
-    return parts[0];
-  }
-
-  private isCRDGroup(manifest: IStageManifest): boolean {
-    return !BUILT_IN_GROUPS.includes(this.apiGroup(manifest));
-  }
-
-  private scopedKind(manifest: IStageManifest): string {
-    if (this.isCRDGroup(manifest)) {
-      return upperFirst(manifest.kind) + '.' + this.apiGroup(manifest);
-    }
-
-    return upperFirst(manifest.kind);
   }
 
   private stageManifestToIManifest(manifest: IStageManifest, account: string): IManifest {
@@ -143,7 +99,7 @@ export class DeployStatus extends React.Component<IExecutionDetailsSectionProps,
                 {manifests.map(manifest => {
                   const uid =
                     manifest.manifest.metadata.uid || KubernetesManifestService.manifestIdentifier(manifest.manifest);
-                  return <ManifestStatus key={uid} manifest={manifest} stage={stage} />;
+                  return <ManifestStatus key={uid} manifest={manifest} account={stage.context.account} />;
                 })}
               </div>
             </div>
