+++ app/scripts/modules/kubernetes/src/pipelines/stages/deployManifest/manifestStatus/ManifestStatus.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { dump } from 'js-yaml';
 
-import { CopyToClipboard, IManifest, ManifestYaml, Overridable } from '@spinnaker/core';
+import { CopyToClipboard, IManifest, ManifestYaml } from '@spinnaker/core';
 
 import { DeployManifestStatusPills } from './DeployStatusPills';
 import { ManifestDetailsLink } from './ManifestDetailsLink';
@@ -10,16 +10,13 @@ import { ManifestEvents } from './ManifestEvents';
 import './ManifestStatus.less';
 
 export interface IManifestStatusProps {
+  account: string;
   manifest: IManifest;
-  stage: any;
 }
 
-@Overridable('kubernetes.v2.pipeline.stages.deployManifest.manifestStatus')
-export class ManifestStatus extends React.Component<IManifestStatusProps> {
-  public render() {
-    const { manifest, stage } = this.props;
-    const { account } = stage.context;
-    return [
+export function ManifestStatus({ account, manifest }: IManifestStatusProps) {
+  return (
+    <>
       <dl className="manifest-status" key="manifest-status">
         <dt>{manifest.manifest.kind}</dt>
         <dd>
@@ -31,7 +28,7 @@ export class ManifestStatus extends React.Component<IManifestStatusProps> {
           &nbsp;
           <DeployManifestStatusPills manifest={manifest} />
         </dd>
-      </dl>,
+      </dl>
       <div className="manifest-support-links" key="manifest-support-links">
         <ManifestYaml
           linkName="YAML"
@@ -39,10 +36,10 @@ export class ManifestStatus extends React.Component<IManifestStatusProps> {
           modalTitle={manifest.manifest.metadata.name}
         />
         <ManifestDetailsLink linkName="Details" manifest={manifest} accountId={account} />
-      </div>,
+      </div>
       <div className="manifest-events pad-left" key="manifest-events">
         <ManifestEvents manifest={manifest} />
-      </div>,
-    ];
-  }
+      </div>
+    </>
+  );
 }
